refactor(parser): define ComprehensionContext type and move constraint comprehension tests
