+++ parser/src/parser/tests/boundary_constraints.rs
@@ -1,8 +1,9 @@
 use super::{build_parse_test, Boundary, BoundaryConstraint, Identifier, Source, SourceSection};
 use crate::{
     ast::{
-        BoundaryStmt::*, Constant, ConstantType::*, Expression::*, MatrixAccess, NamedTraceAccess,
-        PublicInput, Variable, VariableType, VectorAccess,
+        BoundaryStmt::*, Constant, ConstantType::*, Expression::*, Iterable, MatrixAccess,
+        NamedTraceAccess, PublicInput, Range, SourceSection::*, Trace, TraceCols, Variable,
+        VariableType, VectorAccess,
     },
     error::{Error, ParseError},
 };
@@ -210,6 +211,187 @@ fn boundary_constraint_with_variables() {
     ])]);
     build_parse_test!(source).expect_ast(expected);
 }
+
+// CONSTRAINT COMPREHENSION
+// ================================================================================================
+
+#[test]
+fn bc_comprehension_one_iterable_identifier() {
+    let source = "
+    trace_columns:
+        main: [a, b, c[4]]
+
+    boundary_constraints:
+        enf x.first = 0 for x in c";
+
+    let expected = Source(vec![
+        Trace(Trace {
+            main_cols: vec![
+                TraceCols::new(Identifier("a".to_string()), 1),
+                TraceCols::new(Identifier("b".to_string()), 1),
+                TraceCols::new(Identifier("c".to_string()), 4),
+            ],
+            aux_cols: vec![],
+        }),
+        BoundaryConstraints(vec![ConstraintComprehension(
+            BoundaryConstraint::new(
+                NamedTraceAccess::new(Identifier("x".to_string()), 0, 0),
+                Boundary::First,
+                Const(0),
+            ),
+            vec![(
+                Identifier("x".to_string()),
+                Iterable::Identifier(Identifier("c".to_string())),
+            )],
+        )]),
+    ]);
+
+    build_parse_test!(source).expect_ast(expected);
+}
+
+#[test]
+fn bc_comprehension_one_iterable_range() {
+    let source = "
+    trace_columns:
+        main: [a, b, c[4]]
+
+    boundary_constraints:
+        enf x.first = 0 for x in (0..4)";
+
+    let expected = Source(vec![
+        Trace(Trace {
+            main_cols: vec![
+                TraceCols::new(Identifier("a".to_string()), 1),
+                TraceCols::new(Identifier("b".to_string()), 1),
+                TraceCols::new(Identifier("c".to_string()), 4),
+            ],
+            aux_cols: vec![],
+        }),
+        BoundaryConstraints(vec![ConstraintComprehension(
+            BoundaryConstraint::new(
+                NamedTraceAccess::new(Identifier("x".to_string()), 0, 0),
+                Boundary::First,
+                Const(0),
+            ),
+            vec![(
+                Identifier("x".to_string()),
+                Iterable::Range(Range::new(0, 4)),
+            )],
+        )]),
+    ]);
+
+    build_parse_test!(source).expect_ast(expected);
+}
+
+#[test]
+fn bc_comprehension_one_iterable_slice() {
+    let source = "
+    trace_columns:
+        main: [a, b, c[4]]
+
+    boundary_constraints:
+        enf x.first = 0 for x in c[1..3]";
+
+    let expected = Source(vec![
+        Trace(Trace {
+            main_cols: vec![
+                TraceCols::new(Identifier("a".to_string()), 1),
+                TraceCols::new(Identifier("b".to_string()), 1),
+                TraceCols::new(Identifier("c".to_string()), 4),
+            ],
+            aux_cols: vec![],
+        }),
+        BoundaryConstraints(vec![ConstraintComprehension(
+            BoundaryConstraint::new(
+                NamedTraceAccess::new(Identifier("x".to_string()), 0, 0),
+                Boundary::First,
+                Const(0),
+            ),
+            vec![(
+                Identifier("x".to_string()),
+                Iterable::Slice(Identifier("c".to_string()), Range::new(1, 3)),
+            )],
+        )]),
+    ]);
+    build_parse_test!(source).expect_ast(expected);
+}
+
+#[test]
+fn bc_comprehension_two_iterable_identifiers() {
+    let source = "
+    trace_columns:
+        main: [a, b, c[4], d[4]]
+
+    boundary_constraints:
+        enf x.first = y for (x, y) in (c, d)";
+
+    let expected = Source(vec![
+        Trace(Trace {
+            main_cols: vec![
+                TraceCols::new(Identifier("a".to_string()), 1),
+                TraceCols::new(Identifier("b".to_string()), 1),
+                TraceCols::new(Identifier("c".to_string()), 4),
+                TraceCols::new(Identifier("d".to_string()), 4),
+            ],
+            aux_cols: vec![],
+        }),
+        BoundaryConstraints(vec![ConstraintComprehension(
+            BoundaryConstraint::new(
+                NamedTraceAccess::new(Identifier("x".to_string()), 0, 0),
+                Boundary::First,
+                Elem(Identifier("y".to_string())),
+            ),
+            vec![
+                (
+                    Identifier("x".to_string()),
+                    Iterable::Identifier(Identifier("c".to_string())),
+                ),
+                (
+                    Identifier("y".to_string()),
+                    Iterable::Identifier(Identifier("d".to_string())),
+                ),
+            ],
+        )]),
+    ]);
+
+    build_parse_test!(source).expect_ast(expected);
+}
+
+// INVALID BOUNDARY CONSTRAINT COMPREHENSION
+// ================================================================================================
+
+#[test]
+fn err_bc_comprehension_one_member_two_iterables() {
+    let source = "
+    trace_columns:
+        main: [a, b, c[4]]
+
+    boundary_constraints:
+        enf a.first = c for c in (c, d)";
+
+    let error = Error::ParseError(ParseError::InvalidConstraintComprehension(
+        "Number of members and iterables must match".to_string(),
+    ));
+    build_parse_test!(source).expect_error(error);
+}
+
+#[test]
+fn err_bc_comprehension_two_members_one_iterables() {
+    let source = "
+    trace_columns:
+        main: [a, b, c[4]]
+
+    boundary_constraints:
+        enf a.first = c + d for (c, d) in c";
+
+    let error = Error::ParseError(ParseError::InvalidConstraintComprehension(
+        "Number of members and iterables must match".to_string(),
+    ));
+    build_parse_test!(source).expect_error(error);
+}
+
+// INVALID BOUNDARY CONSTRAINTS
+// ================================================================================================
 
 #[test]
 fn err_invalid_variable() {
