+++ parser/src/parser/tests/integrity_constraints.rs
@@ -1,8 +1,10 @@
 use super::{build_parse_test, Identifier, IntegrityConstraint, Source, SourceSection};
 use crate::{
     ast::{
-        Constant, ConstantType::*, ConstraintType, Expression::*, IndexedTraceAccess,
-        IntegrityStmt::*, MatrixAccess, NamedTraceAccess, Variable, VariableType, VectorAccess,
+        Constant, ConstantType::*, ConstraintType, EvaluatorFunction, EvaluatorFunctionCall,
+        Expression::*, IndexedTraceAccess, IntegrityStmt::*, Iterable, MatrixAccess,
+        NamedTraceAccess, Range, SourceSection::*, Trace, TraceCols, Variable, VariableType,
+        VectorAccess,
     },
     error::{Error, ParseError},
 };
@@ -26,13 +28,6 @@ fn integrity_constraints() {
         None,
     )])]);
     build_parse_test!(source).expect_ast(expected);
-}
-
-#[test]
-fn integrity_constraints_invalid() {
-    let source = "integrity_constraints:
-        enf clk' = clk = 1";
-    build_parse_test!(source).expect_unrecognized_token();
 }
 
 #[test]
@@ -71,7 +66,7 @@ fn multiple_integrity_constraints() {
 }
 
 #[test]
-fn integrity_constraint_with_periodic_col() {
+fn ic_with_periodic_col() {
     let source = "
     integrity_constraints:
         enf k0 + b = 0";
@@ -89,7 +84,7 @@ fn integrity_constraint_with_periodic_col() {
 }
 
 #[test]
-fn integrity_constraint_with_random_value() {
+fn ic_with_random_value() {
     let source = "
     integrity_constraints:
         enf a + $rand[1] = 0";
@@ -107,7 +102,7 @@ fn integrity_constraint_with_random_value() {
 }
 
 #[test]
-fn integrity_constraint_with_constants() {
+fn ic_with_constants() {
     let source = "
         const A = 0
         const B = [0, 1]
@@ -149,7 +144,7 @@ fn integrity_constraint_with_constants() {
 }
 
 #[test]
-fn integrity_constraint_with_variables() {
+fn ic_with_variables() {
     let source = "
     integrity_constraints:
         let a = 2^2
@@ -215,7 +210,7 @@ fn integrity_constraint_with_variables() {
 }
 
 #[test]
-fn integrity_constraint_with_indexed_trace_access() {
+fn ic_with_indexed_trace_access() {
     let source = "
     integrity_constraints:
         enf $main[0]' = $main[1] + 1
@@ -245,6 +240,278 @@ fn integrity_constraint_with_indexed_trace_access() {
     build_parse_test!(source).expect_ast(expected);
 }
 
+// CONSTRAINT COMPREHENSION
+// ================================================================================================
+
+#[test]
+fn ic_comprehension_one_iterable_identifier() {
+    let source = "
+    trace_columns:
+        main: [a, b, c[4]]
+
+    integrity_constraints:
+        enf x = a + b for x in c";
+
+    let expected = Source(vec![
+        Trace(Trace {
+            main_cols: vec![
+                TraceCols::new(Identifier("a".to_string()), 1),
+                TraceCols::new(Identifier("b".to_string()), 1),
+                TraceCols::new(Identifier("c".to_string()), 4),
+            ],
+            aux_cols: vec![],
+        }),
+        IntegrityConstraints(vec![ConstraintComprehension(
+            ConstraintType::Inline(IntegrityConstraint::new(
+                Elem(Identifier("x".to_string())),
+                Add(
+                    Box::new(Elem(Identifier("a".to_string()))),
+                    Box::new(Elem(Identifier("b".to_string()))),
+                ),
+            )),
+            None,
+            vec![(
+                Identifier("x".to_string()),
+                Iterable::Identifier(Identifier("c".to_string())),
+            )],
+        )]),
+    ]);
+    build_parse_test!(source).expect_ast(expected);
+}
+
+#[test]
+fn ic_comprehension_one_iterable_range() {
+    let source = "
+    trace_columns:
+        main: [a, b, c[4]]
+    
+    integrity_constraints:
+        enf x = a + b for x in (1..4)";
+
+    let expected = Source(vec![
+        Trace(Trace {
+            main_cols: vec![
+                TraceCols::new(Identifier("a".to_string()), 1),
+                TraceCols::new(Identifier("b".to_string()), 1),
+                TraceCols::new(Identifier("c".to_string()), 4),
+            ],
+            aux_cols: vec![],
+        }),
+        IntegrityConstraints(vec![ConstraintComprehension(
+            ConstraintType::Inline(IntegrityConstraint::new(
+                Elem(Identifier("x".to_string())),
+                Add(
+                    Box::new(Elem(Identifier("a".to_string()))),
+                    Box::new(Elem(Identifier("b".to_string()))),
+                ),
+            )),
+            None,
+            vec![(
+                Identifier("x".to_string()),
+                Iterable::Range(Range::new(1, 4)),
+            )],
+        )]),
+    ]);
+    build_parse_test!(source).expect_ast(expected);
+}
+
+#[test]
+fn ic_comprehension_with_selectors() {
+    let source = "
+    trace_columns:
+        main: [s[2], a, b, c[4]]
+
+    integrity_constraints:
+        enf x = a + b when s[0] & s[1] for x in c";
+
+    let expected = Source(vec![
+        Trace(Trace {
+            main_cols: vec![
+                TraceCols::new(Identifier("s".to_string()), 2),
+                TraceCols::new(Identifier("a".to_string()), 1),
+                TraceCols::new(Identifier("b".to_string()), 1),
+                TraceCols::new(Identifier("c".to_string()), 4),
+            ],
+            aux_cols: vec![],
+        }),
+        IntegrityConstraints(vec![ConstraintComprehension(
+            ConstraintType::Inline(IntegrityConstraint::new(
+                Elem(Identifier("x".to_string())),
+                Add(
+                    Box::new(Elem(Identifier("a".to_string()))),
+                    Box::new(Elem(Identifier("b".to_string()))),
+                ),
+            )),
+            Some(Mul(
+                Box::new(VectorAccess(VectorAccess::new(
+                    Identifier("s".to_string()),
+                    0,
+                ))),
+                Box::new(VectorAccess(VectorAccess::new(
+                    Identifier("s".to_string()),
+                    1,
+                ))),
+            )),
+            vec![(
+                Identifier("x".to_string()),
+                Iterable::Identifier(Identifier("c".to_string())),
+            )],
+        )]),
+    ]);
+    build_parse_test!(source).expect_ast(expected);
+}
+
+#[test]
+fn ic_comprehension_with_evaluator_call() {
+    let source = "
+    ev is_binary(main: [x]):
+        enf x^2 = x
+
+    trace_columns:
+        main: [a, b, c[4], d[4]]
+
+    integrity_constraints:
+        enf is_binary([x]) for x in c";
+
+    let expected = Source(vec![
+        EvaluatorFunction(EvaluatorFunction::new(
+            Identifier("is_binary".to_string()),
+            vec![TraceCols::new(Identifier("x".to_string()), 1)],
+            Vec::new(),
+            vec![Constraint(
+                ConstraintType::Inline(IntegrityConstraint::new(
+                    Exp(
+                        Box::new(Elem(Identifier("x".to_string()))),
+                        Box::new(Const(2)),
+                    ),
+                    Elem(Identifier("x".to_string())),
+                )),
+                None,
+            )],
+        )),
+        Trace(Trace {
+            main_cols: vec![
+                TraceCols::new(Identifier("a".to_string()), 1),
+                TraceCols::new(Identifier("b".to_string()), 1),
+                TraceCols::new(Identifier("c".to_string()), 4),
+                TraceCols::new(Identifier("d".to_string()), 4),
+            ],
+            aux_cols: vec![],
+        }),
+        IntegrityConstraints(vec![ConstraintComprehension(
+            ConstraintType::Evaluator(EvaluatorFunctionCall::new(
+                Identifier("is_binary".to_string()),
+                vec![vec![TraceCols::new(Identifier("x".to_string()), 1)]],
+            )),
+            None,
+            vec![(
+                Identifier("x".to_string()),
+                Iterable::Identifier(Identifier("c".to_string())),
+            )],
+        )]),
+    ]);
+    build_parse_test!(source).expect_ast(expected);
+}
+
+#[test]
+fn ic_comprehension_with_evaluator_and_selectors() {
+    let source = "
+    ev is_binary(main: [x]):
+        enf x^2 = x
+
+    trace_columns:
+        main: [s[2], a, b, c[4], d[4]]
+
+    integrity_constraints:
+        enf is_binary([x]) when s[0] & s[1] for x in c";
+
+    let expected = Source(vec![
+        EvaluatorFunction(EvaluatorFunction::new(
+            Identifier("is_binary".to_string()),
+            vec![TraceCols::new(Identifier("x".to_string()), 1)],
+            Vec::new(),
+            vec![Constraint(
+                ConstraintType::Inline(IntegrityConstraint::new(
+                    Exp(
+                        Box::new(Elem(Identifier("x".to_string()))),
+                        Box::new(Const(2)),
+                    ),
+                    Elem(Identifier("x".to_string())),
+                )),
+                None,
+            )],
+        )),
+        Trace(Trace {
+            main_cols: vec![
+                TraceCols::new(Identifier("s".to_string()), 2),
+                TraceCols::new(Identifier("a".to_string()), 1),
+                TraceCols::new(Identifier("b".to_string()), 1),
+                TraceCols::new(Identifier("c".to_string()), 4),
+                TraceCols::new(Identifier("d".to_string()), 4),
+            ],
+            aux_cols: vec![],
+        }),
+        IntegrityConstraints(vec![ConstraintComprehension(
+            ConstraintType::Evaluator(EvaluatorFunctionCall::new(
+                Identifier("is_binary".to_string()),
+                vec![vec![TraceCols::new(Identifier("x".to_string()), 1)]],
+            )),
+            Some(Mul(
+                Box::new(VectorAccess(VectorAccess::new(
+                    Identifier("s".to_string()),
+                    0,
+                ))),
+                Box::new(VectorAccess(VectorAccess::new(
+                    Identifier("s".to_string()),
+                    1,
+                ))),
+            )),
+            vec![(
+                Identifier("x".to_string()),
+                Iterable::Identifier(Identifier("c".to_string())),
+            )],
+        )]),
+    ]);
+
+    build_parse_test!(source).expect_ast(expected);
+}
+
+// INVALID INTEGRITY CONSTRAINT COMPREHENSION
+// ================================================================================================
+
+#[test]
+fn err_ic_comprehension_one_member_two_iterables() {
+    let source = "
+    trace_columns:
+        main: [a, b, c[4]]
+
+    integrity_constraints:
+        enf a = c for c in (c, d)";
+
+    let error = Error::ParseError(ParseError::InvalidConstraintComprehension(
+        "Number of members and iterables must match".to_string(),
+    ));
+    build_parse_test!(source).expect_error(error);
+}
+
+#[test]
+fn err_ic_comprehension_two_members_one_iterable() {
+    let source = "
+    trace_columns:
+        main: [a, b, c[4]]
+
+    integrity_constraints:
+        enf a = c + d for (c, d) in c";
+
+    let error = Error::ParseError(ParseError::InvalidConstraintComprehension(
+        "Number of members and iterables must match".to_string(),
+    ));
+    build_parse_test!(source).expect_error(error);
+}
+
+// INVALID INTEGRITY CONSTRAINTS
+// ================================================================================================
+
 #[test]
 fn err_missing_integrity_constraint() {
     let source = "
@@ -258,8 +525,12 @@ fn err_missing_integrity_constraint() {
     build_parse_test!(source).expect_error(error);
 }
 
-// UNRECOGNIZED TOKEN ERRORS
-// ================================================================================================
+#[test]
+fn ic_invalid() {
+    let source = "integrity_constraints:
+        enf clk' = clk = 1";
+    build_parse_test!(source).expect_unrecognized_token();
+}
 
 #[test]
 fn error_invalid_next_usage() {
