+++ parser/src/parser/tests/mod.rs
@@ -12,7 +12,6 @@ mod arithmetic_ops;
 mod boundary_constraints;
 mod comments;
 mod constants;
-mod constraint_comprehension;
 mod evaluator_functions;
 mod identifiers;
 mod integrity_constraints;
